Translate NPC names
This commit will change the language files to contain the entity ids and
their translated names. To main.py a few lines were added that check if
a translated name is available. If so, that name will be applied instead
of the default name that is defined in units.dat

close #15
also look at #4