Prevent the player from accessing negative coordinates
Change the check if a tile is walkable or not. The previous assumption
included a try-except-block, but that only managed to prevent the player
from accessing positive coordinates that were not on the level and
didn't take into account that negative indexes on a list access it from
the end. Now there is an if-statement that fixes the bug.

Fix #16